Fix error C2398

+++ engine/src/Graphics/Desktop/OpenGLFrameBuffer.cpp
@@ -121,7 +121,8 @@ namespace robot2D::priv {
     }
 
     void OpenGLFrameBuffer::Resize(const vec2u& newSize) {
-        m_specification.size = {newSize.x, newSize.y};
+        m_specification.size = {static_cast<int>(newSize.x),
+                                static_cast<int>(newSize.y)};
         Invalidate();
     }
 }
